+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Device Client aborts on Red Hat Enterprise Linux ARM](https://github.com/awslabs/aws-iot-device-client/issues/87)
 - [Device Client crashes due to SIGSEGV on Ubuntu x86](https://github.com/awslabs/aws-iot-device-client/issues/86)
 
+## Unreleased
+
+### Fixed
+- Resolved issue where missing pub sub sample files (publish file and subscribe file) would cause a seg fault
+
 ## [1.1.X] - 2021-04-14
 ### Added
 - SDK logging now configurable via CLI or JSON